Initial HalResponseFactory test structure and class
Provides the initial test expectation methods, and methods for creation
of a standard resource and its JSON and XML representations for
comparison purposes. An initial test is written testing that the factory
will create a JSON response if no Accept header is present (which is
analogous to `*.*`, and thus would match `application/json` first).