Implemented collection resource generator strategies
This one required an update to the Strategy interface: instead of
passing each of the hydrators manager, link generator, and, eventually,
the metadata map, to each strategy, instead the metadata map now passes
_itself_ to strategies, along with the instance to create a resource
for, the metadata that matches that instance, and the request instance.

I updated the URL- and route-based resource strategies to accept the new
arguments, and pull what was needed.

A new trait, ExtractCollection, now has the majority of logic for
creating resources representing collections. It forks based on whether a
paginator or a standard iterator was provided, and then creates:

- default data for the collection resource, including `_total_items` in
  both, and `_page` and `_page_count` for paginated sets.
- links. If the metadata does not include a valid pagination parameter
  type (must be one of "query" or "parameter"), and is not a paginator
  instance, it generates only a self relational link. Otherwise, it will
  generate a self link, and, if more than one page is present,
  potentially each of first, prev, next, and last.
- embedded resources. The key used is the collection relation provided
  in the metadata, and the trait iterates over each member, passing it
  to the resource generator's `createResource()` method.

The trait defines two abstract methods, `generateLinkForPage()` and
`generateSelfLink()`, which its logic will call in order to generate
links as needed.

Two implementations, `RouteBasedCollectionStrategy` and
`UrlBasedCollectionStrategy`, compose the trait, define the methods
required by the trait, and then delegate to `extractCollection()` to
generate the collection resource itself.