Provides working resource generator
Implements the `ResourceGenerator`, which is a strategy-based class that
will map a given metadata type to a specific resource generator
strategy. Each has logic for generating a resource with its links,
whether the resource is standalone or composes a collection.

Currently, no logic for handling resource elements that represent other
known resource types; that will come in a later commit.