Allows adding route params and query string args to RouteBased metadata after-the-fact
When dealing with route-based metadata, we may be dealing with
sub-resources; in such cases, the route parameters may be derived from
the request, and we will want to inject them at run-time.

When dealing with collections, the query string arguments may indicate
things such as searches, sort directions, sort columns, filters, limits,
etc.; these will be derived from the request, and need to be injected at
run-time.