feat: Allow Doctrine paginators in collections
This patch adds the method `extractDoctrinePaginator()` to the
`ExtractCollectionTrait`, and modifies `extractCollection()` to detect
Doctrine paginators and invoke the new method when detected.

The method pulls the query from the paginator instance to determine how
many results per page are expected, and runs a count against the
paginator to determine the total number of records possible. These are
then used to create pagination relational links and the various data
pieces we inject in the collection resource (total items, total pages,
etc.)