refactor: extract common pagination routines
Extracts two methods, `createPaginatedCollectionResource()` and
`createCollectionResource()` from the bodies of each of
`extractPaginator()` and `extractDoctrinePaginator()`. The first method
determines if we have a known pagination type; if not, it simply
generates a self relational link and creates a HAL collection by
iterating the paginator. If it does know the pagination type, it will
generate any necessary pagination links before iterating the collection
to produce the resource.

These changes simplify the logic of `extractPaginator()` and
`extractDoctrinePaginator()`, allowing them to focus primarily on how
they determine the required pagination information.